Move logs command to the new command format
